+++ components/staff/DeleteStaffBtn.jsx
@@ -6,7 +6,7 @@ import Spinner from '../Spinner';
 import { GET_STAFF } from './staffQueries';
 import style from '@/styles/register.module.css';
 
-const DeleteStaffBtn = () => {
+const DeleteStaffBtn = ({staff}) => {
   const router = useRouter();
   const { id } = router.query;
   const [deleteStaff, { loading, error }] = useMutation(DELETE_STAFF, {
@@ -17,19 +17,23 @@ const DeleteStaffBtn = () => {
       toast.error(error.message);
     },
     onCompleted: () => {
-      toast.warning('Student deleted successfully');
+      toast.warning(`${staff.firstName}'s profile deleted`);
       router.push('/register');
     },
     refetchQueries: [{ query: GET_STAFF }],
   });
+   const confirmDelete = () => {
+     const deleteProfle = window.confirm(
+       `Are you sure you want to delete ${staff.firstName}'s profile `
+     );
+     deleteProfle && deleteStaff();
+   };
   return (
     <div>
       <button
         type='button'
         className={`${style.btn} ${style.btnCancel} `}
-        onClick={() => {
-          deleteStaff();
-        }}
+        onClick={confirmDelete}
       >
         {loading ? <Spinner /> : <p className='btn btn-secondary'>delete</p>}
       </button>
